requirements.txt
tgcrypto
py-tgcalls
youtube_dl
ffmpeg-python
python-dotenv
youtube_search
youtube_search_python
git+https://github.com/pyrogram/pyrogram@master
